Plot read coverage graphs of contigs classified as Prophage-like, Sloping, or HighCovNoPattern

Plot the read coverages of a contig and its associated pattern-match for Prophage-like, Sloping and HighCovNoPattern classifications. Returns a list of ggplot objects.

Usage

plotTrIdentResults(
  VLPpileup,
  WCpileup,
  TrIdentResults,
  matchScoreFilter,
  saveFilesTo
)

Arguments

VLPpileup

VLP-fraction pileup file.

WCpileup

Whole-community pileup file.

TrIdentResults

Output from `TrIdentClassifier()`.

matchScoreFilter

Optional, Filter plots using the normalized pattern match-scores. A suggested filtering threshold is provided by `TrIdentClassifier()` if `suggFiltThresh=TRUE`.

saveFilesTo

Optional, Provide a path to the directory you wish to save output to. A folder will be made within the provided directory to store results.

Value

Large list containing ggplot objects

Examples

patternMatches <- plotTrIdentResults(
  VLPpileup = VLPFractionSamplePileup,
  WCpileup = WholeCommunitySamplePileup,
  TrIdentResults = TrIdentSampleOutput
)